Plotting circle encoded text
First result of studying necklaces, uses matplotlib to graph the text provided onto a circle in a similar manner to how dnd spells are encoded using the Spellwriting Guide by GorillaOfDestiny. Generation of necklaces to match the alphabet stops after matching everything to prevent lengthy generation times with longer words or text. An encryption feature is also available that shifts the order of the points on the graph and provides a key, greatly obstructing attempts to read the text.