Circle encoding program update for data transfer
With the new circle graphing program you are able to send the binary generated in the output. Anyone who receives this binary data can use it as input to recreate the text from the original. This may also serve as a method for data encryption, although a future program expanding upon that is planned.