Version 2 for encoding data using necklace gen
Trimmed some code, simplified some code, added ability to encode text without splitting, added variable assignments near beginning of file to aid in encrypting data. 
Unsplit data allows for encrypting code, this is handled differently now, lacking a header for word sections and relying of the new flags header and the encryption variables. 
Version 3 is planned to allow for encryption from/to a file, this version was a needed middle step.