Necklace encryption imports key generation.
The key generation file now provides functions for the encryption file to use. Configuration is at the top of the encryption file and can be altered. As before, a, b, c, and d all alter the behavior of the generation program. Try to keep the total length of abcd around or below 60 to leave room for all parts of the key. The rVal variable alters the randomness of the key, if you are experiencing errors when using longer abcd try lowering the rVal. The pKey is the private key. This can be any value but must match between sender and receiver in order to retrieve all necessary information from the public key portion of encrypted messages. 

To use: 
Run necklace_encryption.py and follow the prompts. If encrypting, you can copy any of the three output options as the encrypted text. 
You can alternatively run from the command line, use -h to see more details. You can read from or save to files using command line arguments.